Add MountainCar-v0 environment
The environment have been discretised by splitting the continuous
space in discrete steps. The previous developed Q-Learning algorithm
has been adapted to the environment. Q-Learning is able to solve
the problem but it is still not able to provide an efficient working
solution. It is necessary to perform some tunings between the
exploitation/exploration criterion and the number of episodes/actions.