Change epsilon rate behaviour
In order to avoid the dependency from random moves, the maximum epsilon rate is
set to 1. Doing this the training phase could enter in some infinite learnt loop.
To avoid this if epsilon is 1 and 10000 actions have been performed in
an episode, epsilon is set to 0.9 to have an "help from randomness".
With this modification the trained network can solve the problem without
the need of epsilon rate in testing phase.